レビュー・コメント投稿のjsを統合
