+++ app/javascript/packs/application.js
@@ -11,7 +11,6 @@ require("../comment_delete")
 require("../canvas")
 require("../form")
 require("../drawing")
-require("../review_post")
 
 
 // Uncomment to copy all static images under ../images to the output folder and reference
